Implemented nonperiodic Stencil within a bath. The original toroidal Stencil is now renamed TStencil. Added config file illustrating axonal propagation. Wave propagator now checks Courant condition.
Fixed bug in outputting interval.
Fixed bug for "Steady" intialization.
Fixed bug for white noise buffer overflow.
Debugging (improved but not solved) axonal time delay bug.
Working configurations for cortex.conf and emirs.conf?

+++ solver.h
@@ -51,7 +51,7 @@ class Solver : public NF
     int ncnt;
     double deltat;
     
-    double t;
+    int t;
     Array<Output> m;
     int start; // time to start of output
     int interval; // output interval
